:truck: [#586] Move SOAP service to their own app
