+++ src/openforms/fixtures/default_admin_index.json
@@ -216,7 +216,7 @@
                 "certificate"
             ],
             [
-                "stuf",
+                "soap",
                 "soapservice"
             ],
             [
