+++ src/stuf/admin.py
@@ -3,7 +3,7 @@ from django.contrib import admin
 from django.forms import PasswordInput
 from django.utils.translation import gettext_lazy as _
 
-from .models import SoapService, StufService
+from .models import StufService
 
 
 class StufServiceAdminAdminForm(forms.ModelForm):
@@ -52,39 +52,3 @@ class StufServiceAdmin(admin.ModelAdmin):
             },
         ),
     )
-
-
-@admin.register(SoapService)
-class SoapServiceAdmin(admin.ModelAdmin):
-    list_display = (
-        "label",
-        "url",
-    )
-    search_fields = (
-        "label",
-        "url",
-    )
-    fieldsets = (
-        (
-            None,
-            {
-                "fields": ("label", "url", "soap_version"),
-            },
-        ),
-        (
-            _("Authentication"),
-            {
-                "fields": [
-                    "endpoint_security",
-                    "user",
-                    "password",
-                    "client_certificate",
-                    "server_certificate",
-                ]
-            },
-        ),
-    )
-
-    class Meta:
-        model = SoapService
-        fields = "__all__"
