+++ src/stuf/tests/factories.py
@@ -2,7 +2,8 @@ from pathlib import Path
 
 import factory
 
-from stuf.models import SoapService, StufService
+from soap.tests.factories import SoapServiceFactory
+from stuf.models import StufService
 
 DATA_DIR = Path(__file__).parent.resolve() / "data"
 
@@ -20,26 +21,6 @@ class CertificateFactory(factory.django.DjangoModelFactory):
     class Params:
         with_private_key = factory.Trait(
             private_key=factory.django.FileField(from_path=str(DATA_DIR / "test.key"))
-        )
-
-
-class SoapServiceFactory(factory.django.DjangoModelFactory):
-    label = factory.Sequence(lambda n: f"soap-service-{n}")
-    url = "http://zaken/soap/"
-
-    class Meta:
-        model = SoapService
-
-    class Params:
-        with_server_cert = factory.Trait(
-            server_certificate=factory.SubFactory(
-                CertificateFactory, public_certificate__filename="server.cert"
-            ),
-        )
-        with_client_cert = factory.Trait(
-            client_certificate=factory.SubFactory(
-                CertificateFactory, public_certificate__filename="client.cert"
-            ),
         )
